Split out readings extraction and filtering methods to new builder class, code cleanup.
