+++ src/main/java/gov/usgs/aqcu/retrieval/TimeSeriesDescriptionListService.java
@@ -3,9 +3,6 @@ package gov.usgs.aqcu.retrieval;
 import java.util.List;
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Map;
-import java.util.HashMap;
-import java.util.Set;
 import java.util.HashSet;
 
 import org.slf4j.Logger;
